feat: apply parallel routes to hottrend
